refactor: simplify namespace
