+++ Models/SpotlightDesktop.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 
-namespace SpotlightGallery.Models.Spotlight
+namespace SpotlightGallery.Models
 {
     public class SpotlightDesktopItemContent
     {
