+++ Models/SpotlightLockscreen.cs
@@ -1,5 +1,4 @@
-﻿using SpotlightGallery.Models.Spotlight;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
